requirements.txt
alembic==1.14.0
amqp==5.2.0
async-timeout==5.0.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
redis==5.2.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.1
XlsxWriter==3.2.0
